oruwan_challengeAI/requirements.txt
absl-py==1.4.0
APScheduler==3.10.1
asgiref==3.7.2
asttokens==2.2.1
astunparse==1.6.3
backcall==0.2.0
cachetools==5.3.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
colorama==0.4.6
comm==0.1.2
cryptography==41.0.1
debugpy==1.6.5
decorator==5.1.1
defusedxml==0.7.1
Django==4.2.2
django-allauth==0.54.0
django-apscheduler==0.6.2
django-cors-headers==4.0.0
django-cron==0.6.0
django-crontab==0.7.1
django-rest-framework==0.1.0
django-webpack-loader==2.0.0
djangorestframework==3.14.0
entrypoints==0.4
executing==1.2.0
filelock==3.12.2
flatbuffers==23.5.26
fsspec==2023.6.0
gast==0.4.0
google-auth==2.20.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.54.2
h5py==3.8.0
huggingface-hub==0.15.1
idna==3.4
ipykernel==6.19.4
ipython==8.8.0
jax==0.4.12
jedi==0.18.2
jupyter_client==7.4.8
jupyter_core==5.1.2
keras==2.12.0
libclang==16.0.0
Markdown==3.4.3
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
ml-dtypes==0.2.0
mysqlclient==2.1.1
nest-asyncio==1.5.6
numpy==1.23.5
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==22.0
pandas==2.0.3
parso==0.8.3
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==2.6.2
prompt-toolkit==3.0.36
protobuf==4.23.3
psutil==5.9.4
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
Pygments==2.14.0
PyJWT==2.7.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2023.3
pywin32==305
PyYAML==6.0
pyzmq==24.0.1
regex==2023.6.3
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
safetensors==0.3.1
scipy==1.10.1
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.2
tensorboard==2.12.3
tensorboard-data-server==0.7.1
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-hub==0.13.0
tensorflow-intel==2.12.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
tokenizers==0.13.3
tornado==6.2
tqdm==4.65.0
traitlets==5.8.0
transformers==4.30.2
typing_extensions==4.6.3
tzdata==2023.3
tzlocal==5.0.1
urllib3==1.26.16
wcwidth==0.2.5
Werkzeug==2.3.6
wrapt==1.14.1
